@@ -12,19 +12,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.2.0",
+    "@angular/animations": "^18.2.1",
     "@angular/cdk": "^14.2.0",
-    "@angular/common": "^14.2.0",
-    "@angular/compiler": "^14.2.0",
-    "@angular/core": "^14.2.0",
-    "@angular/elements": "^14.2.0",
-    "@angular/forms": "^14.2.0",
+    "@angular/common": "^18.2.1",
+    "@angular/compiler": "^18.2.1",
+    "@angular/core": "^18.2.1",
+    "@angular/elements": "^18.2.1",
+    "@angular/forms": "^18.2.1",
     "@angular/google-maps": "^14.2.0",
-    "@angular/localize": "^14.2.0",
+    "@angular/localize": "^18.2.1",
     "@angular/material": "^14.2.0",
-    "@angular/platform-browser": "^14.2.0",
-    "@angular/platform-browser-dynamic": "^14.2.0",
-    "@angular/router": "^14.2.0",
+    "@angular/platform-browser": "^18.2.1",
+    "@angular/platform-browser-dynamic": "^18.2.1",
+    "@angular/router": "^18.2.1",
     "@ng-bootstrap/ng-bootstrap": "^13.0.0",
     "@popperjs/core": "^2.11.4",
     "bootstrap": "4.6.1",
@@ -35,6 +35,7 @@
     "ngx-sweetalert2": "^0.2.7",
     "ngx-toastr": "14.2.2",
     "nouislider": "15.5.1",
+    "pdf-lib": "^1.17.1",
     "rxjs": "~7.5.0",
     "sweetalert2": "^11.6.13",
     "web-animations-js": "2.3.2",
@@ -42,12 +43,12 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^18.0.7",
-    "@angular/cli": "~14.2.7",
-    "@angular/compiler-cli": "^14.2.0",
-    "@angular/language-service": "14.2.0",
+    "@angular/cli": "~18.2.1",
+    "@angular/compiler-cli": "^18.2.1",
+    "@angular/language-service": "18.2.1",
     "@types/jasmine": "~4.0.0",
     "@types/jasminewd2": "~2.0.10",
-    "@types/node": "^17.0.21",
+    "@types/node": "^22.5.1",
     "codelyzer": "^0.0.28",
     "jasmine-core": "~4.4.0",
     "jasmine-spec-reporter": "~7.0.0",
@@ -57,8 +58,8 @@
     "karma-coverage-istanbul-reporter": "~3.0.3",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
-    "protractor": "^3.3.0",
+    "protractor": "^7.0.0",
     "ts-node": "~10.9.1",
-    "typescript": "~4.7.2"
+    "typescript": "~5.4.5"
   }
 }
