@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { RouterModule } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
@@ -59,66 +59,60 @@ import { ParametrosComponent } from 'src/app/pages/parametros/parametros.compone
 
 
 
-@NgModule({
-  imports: [
-    CommonModule,
-    RouterModule.forChild(AdminLayoutRoutes),
-    FormsModule,
-    HttpClientModule,
-    NgbModule,
-    ClipboardModule,
-    NgbModalModule,
-    MatTabsModule,
-  ],
-  declarations: [
-    DashboardComponent,
-    UserProfileComponent,
-    TablesComponent,
-    IconsComponent,
-    MapsComponent,
-    BancoComponent,
-    PersistBancoComponent,
-    MoedaComponent,
-    PersistMoedaComponent,
-    ContaComponent,
-    PersistContaComponent,
-    ActividadeComponent,
-    PersistActividadeComponent,
-    CoordenadorComponent,
-    PersistCoordenadorComponent,
-    BeneficiarioComponent,
-    PersistBeneficiarioComponent,
-    UnidadeOrganicaComponent,
-    PersistUnidadeOrganicaComponent,
-    PaisComponent,
-    PersistPaisComponent,
-    RubricaEstadoComponent,
-    PersistRubricaEstadoComponent,
-    FinanciadorComponent,
-    PersistFinanciadorComponent,
-    FinanciamentoComponent,
-    PersistFinanciamentoComponent,
-    ProjectoComponent,
-    PersistProjectoComponent,
-    ShowBeneficiarioComponent,
-    ShowFinanciadorComponent,
-    ShowProjectoComponent,
-    ShowRubricaProjectoComponent,
-    SubRubricaComponent,
-    RequisicaoComponent,
-    PersistRequisicaoRubrica,
-    RequisicaoRubricaComponent,
-    CambioComponent,
-    PersistCambioComponent,
-    FornecedorComponent,
-    PersistFornecedorComponent,
-    PersistAssinanteComponent,
-    AssinanteComponent,
-    SubProjectoComponent,
-    PersistSubProjectoComponent,
-    ParametrosComponent,
-  ],
-})
+@NgModule({ declarations: [
+        DashboardComponent,
+        UserProfileComponent,
+        TablesComponent,
+        IconsComponent,
+        MapsComponent,
+        BancoComponent,
+        PersistBancoComponent,
+        MoedaComponent,
+        PersistMoedaComponent,
+        ContaComponent,
+        PersistContaComponent,
+        ActividadeComponent,
+        PersistActividadeComponent,
+        CoordenadorComponent,
+        PersistCoordenadorComponent,
+        BeneficiarioComponent,
+        PersistBeneficiarioComponent,
+        UnidadeOrganicaComponent,
+        PersistUnidadeOrganicaComponent,
+        PaisComponent,
+        PersistPaisComponent,
+        RubricaEstadoComponent,
+        PersistRubricaEstadoComponent,
+        FinanciadorComponent,
+        PersistFinanciadorComponent,
+        FinanciamentoComponent,
+        PersistFinanciamentoComponent,
+        ProjectoComponent,
+        PersistProjectoComponent,
+        ShowBeneficiarioComponent,
+        ShowFinanciadorComponent,
+        ShowProjectoComponent,
+        ShowRubricaProjectoComponent,
+        SubRubricaComponent,
+        RequisicaoComponent,
+        PersistRequisicaoRubrica,
+        RequisicaoRubricaComponent,
+        CambioComponent,
+        PersistCambioComponent,
+        FornecedorComponent,
+        PersistFornecedorComponent,
+        PersistAssinanteComponent,
+        AssinanteComponent,
+        SubProjectoComponent,
+        PersistSubProjectoComponent,
+        ParametrosComponent,
+    ], imports: [CommonModule,
+        RouterModule.forChild(AdminLayoutRoutes),
+        FormsModule,
+        NgbModule,
+        ClipboardModule,
+        NgbModalModule,
+        MatTabsModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 
 export class AdminLayoutModule {
 }
