@@ -1,5 +1,5 @@
 
-import {HttpClient} from "@angular/common/http";
+import { HttpClient } from "@angular/common/http";
 import {environment} from "../../../environments/environment";
 import {Observable} from "rxjs";
 import {Cambio} from "../../models/cambio/cambio";
