@@ -1,5 +1,5 @@
 import {Injectable} from "@angular/core";
-import {HttpClient, HttpErrorResponse} from "@angular/common/http";
+import { HttpClient, HttpErrorResponse } from "@angular/common/http";
 import {environment} from "../../../environments/environment";
 import {catchError, Observable, throwError} from "rxjs";
 import {Financiador} from "../../models/financiador/financiador";
