@@ -1,6 +1,6 @@
 import {Injectable} from "@angular/core";
 import {environment} from "../../../environments/environment";
-import {HttpClient, HttpErrorResponse} from "@angular/common/http";
+import { HttpClient, HttpErrorResponse } from "@angular/common/http";
 import {SubRubrica} from "../../models/subRubrica/subRubrica";
 import {catchError, Observable, throwError} from "rxjs";
 
