@@ -1,6 +1,6 @@
 import {Injectable} from "@angular/core";
 import {environment} from "../../../environments/environment";
-import {HttpClient} from "@angular/common/http";
+import { HttpClient } from "@angular/common/http";
 import {UnidadeOrganica} from "../../models/unidadeOrganica/unidadeOrganica";
 import {Observable} from "rxjs";
 import {Conta} from "../../models/conta/conta";
